+++ src/backend/base/langflow/interface/embeddings/base.py
@@ -1,13 +1,13 @@
 from typing import Dict, List, Optional, Type
 
-from loguru import logger
-
 from langflow.interface.base import LangChainTypeCreator
 from langflow.interface.custom_lists import embedding_type_to_cls_dict
+from langflow.interface.utils import build_template_from_class
+from langflow.services.deps import get_settings_service
+from langflow.interface.utils import build_template_from_class
 from langflow.services.deps import get_settings_service
 from langflow.template.frontend_node.base import FrontendNode
 from langflow.template.frontend_node.embeddings import EmbeddingFrontendNode
-from langflow.utils.util import build_template_from_class
 
 
 class EmbeddingCreator(LangChainTypeCreator):
