+++ src/backend/base/langflow/interface/output_parsers/base.py
@@ -1,13 +1,13 @@
 from typing import ClassVar, Dict, List, Optional, Type
 
 from langchain import output_parsers
-from loguru import logger
-
 from langflow.interface.base import LangChainTypeCreator
 from langflow.interface.importing.utils import import_class
+from langflow.interface.utils import build_template_from_class
 from langflow.services.deps import get_settings_service
 from langflow.template.frontend_node.output_parsers import OutputParserFrontendNode
-from langflow.utils.util import build_template_from_class, build_template_from_method
+from langflow.utils.util import build_template_from_method
+from loguru import logger
 
 
 class OutputParserCreator(LangChainTypeCreator):
