+++ src/backend/base/langflow/interface/prompts/base.py
@@ -1,14 +1,14 @@
 from typing import Dict, List, Optional, Type
 
 from langchain import prompts
-from loguru import logger
-
 from langflow.interface.base import LangChainTypeCreator
 from langflow.interface.importing.utils import import_class
+from langflow.interface.utils import build_template_from_class
+from langflow.legacy_custom.customs import get_custom_nodes
+from langflow.interface.utils import build_template_from_class
 from langflow.legacy_custom.customs import get_custom_nodes
 from langflow.services.deps import get_settings_service
 from langflow.template.frontend_node.prompts import PromptFrontendNode
-from langflow.utils.util import build_template_from_class
 
 
 class PromptCreator(LangChainTypeCreator):
