+++ src/backend/base/langflow/interface/retrievers/base.py
@@ -1,13 +1,13 @@
 from typing import Any, ClassVar, Dict, List, Optional, Type
 
 from langchain_community import retrievers
-from loguru import logger
-
 from langflow.interface.base import LangChainTypeCreator
 from langflow.interface.importing.utils import import_class
+from langflow.interface.utils import build_template_from_class
 from langflow.services.deps import get_settings_service
 from langflow.template.frontend_node.retrievers import RetrieverFrontendNode
-from langflow.utils.util import build_template_from_class, build_template_from_method
+from langflow.utils.util import build_template_from_method
+from loguru import logger
 
 
 class RetrieverCreator(LangChainTypeCreator):
