+++ src/backend/base/langflow/interface/wrappers/base.py
@@ -1,10 +1,9 @@
 from typing import Dict, List, Optional
 
 from langchain_community.utilities import requests
-from loguru import logger
-
 from langflow.interface.base import LangChainTypeCreator
-from langflow.utils.util import build_template_from_class
+from langflow.interface.utils import build_template_from_class
+from loguru import logger
 
 
 class WrapperCreator(LangChainTypeCreator):
